third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-lab-expected.txt

This is a testharness.js-based test.
Found 92 tests; 12 PASS, 80 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Property color value 'lab(0 0 0)' assert_equals: expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL Property color value 'lab(0 0 0 / 1)' assert_equals: expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL Property color value 'lab(0 0 0 / 0.5)' assert_equals: expected "lab(0 0 0 / 0.5)" but got "lab(0% 0 0 / 0.5)"
FAIL Property color value 'lab(20 0 10/0.5)' assert_equals: expected "lab(20 0 10 / 0.5)" but got "lab(20% 0 10 / 0.5)"
FAIL Property color value 'lab(20 0 10/50%)' assert_equals: expected "lab(20 0 10 / 0.5)" but got "lab(20% 0 10 / 0.5)"
FAIL Property color value 'lab(400 0 10/50%)' assert_equals: expected "lab(400 0 10 / 0.5)" but got "lab(400% 0 10 / 0.5)"
FAIL Property color value 'lab(50 -160 160)' assert_equals: expected "lab(50 -160 160)" but got "lab(50% -160 160)"
FAIL Property color value 'lab(50 -200 200)' assert_equals: expected "lab(50 -200 200)" but got "lab(50% -200 200)"
FAIL Property color value 'lab(0 0 0 / -10%)' assert_equals: expected "lab(0 0 0 / 0)" but got "lab(0% 0 0 / 0)"
FAIL Property color value 'lab(0 0 0 / 110%)' assert_equals: expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL Property color value 'lab(0 0 0 / 300%)' assert_equals: expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL Property color value 'lab(-40 0 0)' assert_equals: expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL Property color value 'lab(50 -20 0)' assert_equals: expected "lab(50 -20 0)" but got "lab(50% -20 0)"
FAIL Property color value 'lab(50 0 -20)' assert_equals: expected "lab(50 0 -20)" but got "lab(50% 0 -20)"
FAIL Property color value 'lab(calc(50 * 3) calc(0.5 - 1) calc(1.5) / calc(-0.5 + 1))' assert_equals: expected "lab(150 -0.5 1.5 / 0.5)" but got "lab(150% -0.5 1.5 / 0.5)"
FAIL Property color value 'lab(calc(-50 * 3) calc(0.5 + 1) calc(-1.5) / calc(-0.5 * 2))' assert_equals: expected "lab(0 1.5 -1.5 / 0)" but got "lab(0% 1.5 -1.5 / 0)"
PASS Property color value 'lab(none none none / none)'
PASS Property color value 'lab(none none none)'
FAIL Property color value 'lab(20 none none / none)' assert_equals: expected "lab(20 none none / none)" but got "lab(20% none none / none)"
PASS Property color value 'lab(none none none / 0.5)'
FAIL Property color value 'lab(0 0 0 / none)' assert_equals: expected "lab(0 0 0 / none)" but got "lab(0% 0 0 / none)"
FAIL Property color value 'oklab(0 0 0)' assert_equals: expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL Property color value 'oklab(0 0 0 / 1)' assert_equals: expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL Property color value 'oklab(0 0 0 / 0.5)' assert_equals: expected "oklab(0 0 0 / 0.5)" but got "oklab(0% 0 0 / 0.5)"
FAIL Property color value 'oklab(20 0 10/0.5)' assert_equals: expected "oklab(20 0 10 / 0.5)" but got "oklab(20% 0 10 / 0.5)"
FAIL Property color value 'oklab(20 0 10/50%)' assert_equals: expected "oklab(20 0 10 / 0.5)" but got "oklab(20% 0 10 / 0.5)"
FAIL Property color value 'oklab(400 0 10/50%)' assert_equals: expected "oklab(400 0 10 / 0.5)" but got "oklab(400% 0 10 / 0.5)"
FAIL Property color value 'oklab(50 -160 160)' assert_equals: expected "oklab(50 -160 160)" but got "oklab(50% -160 160)"
FAIL Property color value 'oklab(50 -200 200)' assert_equals: expected "oklab(50 -200 200)" but got "oklab(50% -200 200)"
FAIL Property color value 'oklab(0 0 0 / -10%)' assert_equals: expected "oklab(0 0 0 / 0)" but got "oklab(0% 0 0 / 0)"
FAIL Property color value 'oklab(0 0 0 / 110%)' assert_equals: expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL Property color value 'oklab(0 0 0 / 300%)' assert_equals: expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL Property color value 'oklab(-40 0 0)' assert_equals: expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL Property color value 'oklab(50 -20 0)' assert_equals: expected "oklab(50 -20 0)" but got "oklab(50% -20 0)"
FAIL Property color value 'oklab(50 0 -20)' assert_equals: expected "oklab(50 0 -20)" but got "oklab(50% 0 -20)"
FAIL Property color value 'oklab(calc(50 * 3) calc(0.5 - 1) calc(1.5) / calc(-0.5 + 1))' assert_equals: expected "oklab(150 -0.5 1.5 / 0.5)" but got "oklab(150% -0.5 1.5 / 0.5)"
FAIL Property color value 'oklab(calc(-50 * 3) calc(0.5 + 1) calc(-1.5) / calc(-0.5 * 2))' assert_equals: expected "oklab(0 1.5 -1.5 / 0)" but got "oklab(0% 1.5 -1.5 / 0)"
PASS Property color value 'oklab(none none none / none)'
PASS Property color value 'oklab(none none none)'
FAIL Property color value 'oklab(20 none none / none)' assert_equals: expected "oklab(20 none none / none)" but got "oklab(20% none none / none)"
PASS Property color value 'oklab(none none none / 0.5)'
FAIL Property color value 'oklab(0 0 0 / none)' assert_equals: expected "oklab(0 0 0 / none)" but got "oklab(0% 0 0 / none)"
FAIL Property color value 'lch(0 0 0deg)' assert_equals: expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL Property color value 'lch(0 0 0deg / 1)' assert_equals: expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL Property color value 'lch(0 0 0deg / 0.5)' assert_equals: expected "lch(0 0 0 / 0.5)" but got "lch(0% 0 0 / 0.5)"
FAIL Property color value 'lch(100 230 0deg / 0.5)' assert_equals: expected "lch(100 230 0 / 0.5)" but got "lch(100% 230 0 / 0.5)"
FAIL Property color value 'lch(20 50 20deg/0.5)' assert_equals: expected "lch(20 50 20 / 0.5)" but got "lch(20% 50 20 / 0.5)"
FAIL Property color value 'lch(20 50 20deg/50%)' assert_equals: expected "lch(20 50 20 / 0.5)" but got "lch(20% 50 20 / 0.5)"
FAIL Property color value 'lch(10 20 20deg / -10%)' assert_equals: expected "lch(10 20 20 / 0)" but got "lch(10% 20 20 / 0)"
FAIL Property color value 'lch(10 20 20deg / 110%)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(10 20 1.28rad)' assert_equals: expected "lch(10 20 73.3386)" but got "lch(10% 20 73.3386)"
FAIL Property color value 'lch(10 20 380deg)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(10 20 -340deg)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(10 20 740deg)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(10 20 -700deg)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(-40 0 0)' assert_equals: expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL Property color value 'lch(20 -20 0)' assert_equals: expected "lch(20 0 0)" but got "lch(20% 0 0)"
FAIL Property color value 'lch(0 0 0 / 0.5)' assert_equals: expected "lch(0 0 0 / 0.5)" but got "lch(0% 0 0 / 0.5)"
FAIL Property color value 'lch(10 20 20 / 110%)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(10 20 -700)' assert_equals: expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL Property color value 'lch(calc(50 * 3) calc(0.5 - 1) calc(20deg * 2) / calc(-0.5 + 1))' assert_equals: expected "lch(150 0 40 / 0.5)" but got "lch(150% 0 40 / 0.5)"
FAIL Property color value 'lch(calc(-50 * 3) calc(0.5 + 1) calc(-20deg * 2) / calc(-0.5 * 2))' assert_equals: expected "lch(0 1.5 320 / 0)" but got "lch(0% 1.5 320 / 0)"
PASS Property color value 'lch(none none none / none)'
PASS Property color value 'lch(none none none)'
FAIL Property color value 'lch(20 none none / none)' assert_equals: expected "lch(20 none none / none)" but got "lch(20% none none / none)"
PASS Property color value 'lch(none none none / 0.5)'
FAIL Property color value 'lch(0 0 0 / none)' assert_equals: expected "lch(0 0 0 / none)" but got "lch(0% 0 0 / none)"
FAIL Property color value 'oklch(0 0 0deg)' assert_equals: expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL Property color value 'oklch(0 0 0deg / 1)' assert_equals: expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL Property color value 'oklch(0 0 0deg / 0.5)' assert_equals: expected "oklch(0 0 0 / 0.5)" but got "oklch(0% 0 0 / 0.5)"
FAIL Property color value 'oklch(100 230 0deg / 0.5)' assert_equals: expected "oklch(100 230 0 / 0.5)" but got "oklch(100% 230 0 / 0.5)"
FAIL Property color value 'oklch(20 50 20deg/0.5)' assert_equals: expected "oklch(20 50 20 / 0.5)" but got "oklch(20% 50 20 / 0.5)"
FAIL Property color value 'oklch(20 50 20deg/50%)' assert_equals: expected "oklch(20 50 20 / 0.5)" but got "oklch(20% 50 20 / 0.5)"
FAIL Property color value 'oklch(10 20 20deg / -10%)' assert_equals: expected "oklch(10 20 20 / 0)" but got "oklch(10% 20 20 / 0)"
FAIL Property color value 'oklch(10 20 20deg / 110%)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(10 20 1.28rad)' assert_equals: expected "oklch(10 20 73.3386)" but got "oklch(10% 20 73.3386)"
FAIL Property color value 'oklch(10 20 380deg)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(10 20 -340deg)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(10 20 740deg)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(10 20 -700deg)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(-40 0 0)' assert_equals: expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL Property color value 'oklch(20 -20 0)' assert_equals: expected "oklch(20 0 0)" but got "oklch(20% 0 0)"
FAIL Property color value 'oklch(0 0 0 / 0.5)' assert_equals: expected "oklch(0 0 0 / 0.5)" but got "oklch(0% 0 0 / 0.5)"
FAIL Property color value 'oklch(10 20 20 / 110%)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(10 20 -700)' assert_equals: expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL Property color value 'oklch(calc(50 * 3) calc(0.5 - 1) calc(20deg * 2) / calc(-0.5 + 1))' assert_equals: expected "oklch(150 0 40 / 0.5)" but got "oklch(150% 0 40 / 0.5)"
FAIL Property color value 'oklch(calc(-50 * 3) calc(0.5 + 1) calc(-20deg * 2) / calc(-0.5 * 2))' assert_equals: expected "oklch(0 1.5 320 / 0)" but got "oklch(0% 1.5 320 / 0)"
PASS Property color value 'oklch(none none none / none)'
PASS Property color value 'oklch(none none none)'
FAIL Property color value 'oklch(20 none none / none)' assert_equals: expected "oklch(20 none none / none)" but got "oklch(20% none none / none)"
PASS Property color value 'oklch(none none none / 0.5)'
FAIL Property color value 'oklch(0 0 0 / none)' assert_equals: expected "oklch(0 0 0 / none)" but got "oklch(0% 0 0 / none)"
Harness: the test ran to completion.
